refactor: template

+++ todolist/views.py
@@ -14,17 +14,18 @@ def index(request):
             title = request.POST["description"] #title
             date = str(request.POST["date"]) #date
             category = request.POST["category_select"] #category
-            content = title + " -- " + date + " " + category #content
-            Todo = TodoList(title=title, content=content, due_date=date, category=Category.objects.get(name=category))
-            Todo.save() #saving the todo 
-            return redirect("/") #reloading the page
+            if category in categories:
+                content = title + " -- " + date + " " + category #content
+                Todo = TodoList(title=title, content=content, due_date=date, category=Category.objects.get(name=category))
+                Todo.save() #saving the todo
+                return redirect("/") #reloading the page
 
         if "taskDelete" in request.POST:
             checkedlist = request.POST["checkedbox"] #checked todos to be deleted
-
-            for todo_id in checkedlist:
-                todo = TodoList.objects.get(id=int(todo_id)) #getting todo id
-                todo.delete() #deleting todo
+            if checkedlist:
+                for todo_id in checkedlist:
+                    todo = TodoList.objects.get(id=int(todo_id)) #getting todo id
+                    todo.delete() #deleting todo
